refactor: Calculator에서 InputView 관련 로직 제거, CalculatorController로 역할 위임
