+++ src/main/java/calculator/model/Calculator.java
@@ -3,19 +3,11 @@ package calculator.model;
 import java.util.List;
 
 import calculator.utils.InputValidator;
-import calculator.view.InputView;
 
 public class Calculator {
     private static final String DEFAULT_NUMBER = "0";
 
-    public int sumAndDivideInput(String input) {
-        if (InputView.hasCustomDelimiterInInput(input)) {
-            return makeSumOfNumbers(InputView.dividesByCustomAndCheckValidate(input));
-        }
-        return makeSumOfNumbers(InputView.divideNumbersByDelimiterAndCheckValidate(input));
-    }
-
-    private int makeSumOfNumbers(List<String> numbers) {
+    public int makeSumOfNumbers(List<String> numbers) {
         return numbers.stream()
                 .mapToInt(num -> Integer.parseInt(sumOfNumber(num)))
                 .sum();
